Refactor tests to two files
